# doc/api.yml
openapi: 3.1.0
info:
  title: Book Management API
  version: 2.0.0
  description: >
    API REST pour la gestion d'une librairie :
    œuvres (books), exemplaires (copies), arrivages (arrivals),
    ventes/réservations (sales) et indicateurs.
servers:
  - url: http://localhost:8080
    description: Développement local
...
paths:
...
  /books/{bookId}:
    parameters:
      - $ref: '#/components/parameters/bookId'
    get:
      tags: [books]
      summary: Détail d'un livre
      operationId: getBook
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [books]
      summary: Met à jour un livre
      operationId: updateBook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBookDTO'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    delete:
      tags: [books]
      summary: Supprime un livre
      operationId: deleteBook
      responses:
        '204':
          description: Supprimé
        '404':
          $ref: '#/components/responses/NotFound'
...
  /sales:
    get:
      tags: [sales]
      summary: Liste des ventes (filtre par plage de dates)
      operationId: listSales
      parameters:
        - name: from
          in: query
          description: Date début (ISO 8601 avec timezone, ex. 2024-01-01T00:00:00Z)
          schema: { type: string, format: date-time }
        - name: to
          in: query
          description: Date fin (ISO 8601 avec timezone)
          schema: { type: string, format: date-time }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Sale'
    post:
      tags: [sales]
      summary: Enregistre une vente / réservation
      operationId: createSale
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSaleDTO'
      responses:
        '201':
          description: Créé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sale'
        '400':
          $ref: '#/components/responses/BadRequest'
  /sales/{saleId}:
    parameters:
      - $ref: '#/components/parameters/saleId'
    get:
      tags: [sales]
      summary: Détail d'une vente
      operationId: getSale
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sale'
        '404':
          $ref: '#/components/responses/NotFound'
  /sales/{saleId}/finalize:
    parameters:
      - $ref: '#/components/parameters/saleId'
    post:
      tags: [sales]
      summary: Finalise (confirme ou rejette) une vente
      operationId: finalizeSale
      parameters:
        - name: confirm
          in: query
          required: true
          schema: { type: boolean }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sale'
        '404':
          $ref: '#/components/responses/NotFound'
  /sales/day-revenue:
    get:
      tags: [sales]
      summary: Chiffre d'affaires du jour ou du jour spécifié (ventes confirmées)
      operationId: getDayRevenue
      parameters:
        - name: t
          required: false
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DayRevenue'
  /sales/month-revenue:
    get:
      tags: [sales]
      summary: Chiffre d'affaires du mois en cours ou du mois spécifié
      operationId: getMonthRevenue
      parameters:
        - name: t
          required: false
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MonthRevenue'
  /sales/top-sellers:
    get:
      tags: [sales]
      summary: Livres les plus vendus
      operationId: getTopSellers
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 10 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TopSellerEntry'
...
components:
  parameters:
...
    bookId:
      name: bookId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    saleId:
      name: saleId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    BadRequest:
      description: Requête invalide
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Ressource introuvable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Conflit (ressource déjà existante)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    SaleStatus:
      type: string
      enum: [PENDING, CONFIRMED, REJECTED]

    # Genre
    Genre:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
...
    Author:
      type: object
      properties:
        id:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
...
    Book:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        isbn:
          type: string
        description:
          type: string
        publishDate:
          type: string
          format: date
        genres:
          type: array
          items:
            $ref: '#/components/schemas/Genre'
        authors:
          type: array
          items:
            $ref: '#/components/schemas/Author'
    CreateBookDTO:
      type: object
      required: [title, isbn, publishDate, genreIds, authorIds]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
        isbn:
          type: string
          pattern: '^(97[89])?[0-9]{9}[0-9X]$'
          example: '9780451524935'
        description:
          type: string
        publishDate:
          type: string
          format: date
        genreIds:
          type: array
          items:
            type: string
            format: uuid
          minItems: 1
        authorIds:
          type: array
          items:
            type: string
            format: uuid
          minItems: 1

    # BookCopy
...
    QuantifiedBookCopyDTO:
      type: object
      required: [bookCopyId, quantity]
      properties:
        bookCopyId:
          type: string
          format: uuid
        quantity:
          type: integer
          minimum: 1
...
    SaleBookCopyResponse:
      type: object
      properties:
        bookCopyId:
          type: string
          format: uuid
        price:
          type: number
          format: double
        quantity:
          type: integer
    Sale:
      type: object
      properties:
        id:
          type: string
          format: uuid
        creationDateTime:
          type: string
          format: date-time
        saleStatus:
          $ref: '#/components/schemas/SaleStatus'
        finalizationDateTime:
          type: string
          format: date-time
          nullable: true
        isReservation:
          type: boolean
        books:
          type: array
          items:
            $ref: '#/components/schemas/SaleBookCopyResponse'
    CreateSaleDTO:
      type: object
      required: [creationDateTime, quantifiedBookCopyList, isReservation]
      properties:
        creationDateTime:
          type: string
          format: date-time
        quantifiedBookCopyList:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/QuantifiedBookCopyDTO'
        isReservation:
          type: boolean

    # Dashboard / Revenue
    DayRevenue:
      type: object
      properties:
        todayRevenue:
          type: number
          format: double
    MonthRevenue:
      type: object
      properties:
        monthlyRevenue:
          type: number
          format: double
    TopSellerEntry:
      type: object
      properties:
        bookId:
          type: string
          format: uuid
        title:
          type: string
        unitsSold:
          type: integer
...
    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        timestamp:
          type: string
          format: date-time